requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
docker-pycreds==0.4.0
gitdb==4.0.9
GitPython==3.1.29
idna==3.4
numpy==1.23.4
pandas==1.5.1
pathlib==1.0.1
pathtools==0.1.2
Pillow==9.3.0
promise==2.3
protobuf==4.21.9
psutil==5.9.4
PyHessian==0.1
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
scipy==1.9.3
sentry-sdk==1.11.0
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
torch==1.8.0+cu111
torchvision==0.9.0+cu111
tqdm==4.64.1
typing-extensions==4.4.0
urllib3==1.26.12
wandb==0.13.5
